requirements.txt
django-pwa==1.0.10